ContactController now has added function to open conversation when clicking on a contact in the friend list. The method will ask the server for a conversation if one does not exist and wait for it to be created.
Conversation controller, minor cleanup. In initialize the name and date will be set according to the data from
 the conversation element.

The main screen is responsible for creating the conversation box and setting the controller of the conversation box pane.

Removed unused conversation left over in client.

Minor refactoring and updates to the UI in general.

Created new conversation handler for incoming conversations from the server.

Fixing path issues in the friendlist handler