Made conversation placeholder work
Now the clients stores their conversations in a folder based on their username.
When a user logs in they get a placeholder for conversations that they are in.
Any placeholder get upgraded to a real conversation and stored when new message is sent to the conversation or they enter it.