Server Testability
Changes introduced to make the Server class more testable

## IsRunningStrategy

Interfaces renamed to make it more general, to be used by both Worker and Server. It is also moved out of the worker package to make it for more general use.

### CheckFieldServerRunningStrategy

Made a strategy for the Server class that checks the Server#isRunning method to return if the server is running.

## Server Socket

Made the main method give ServerSocket to the start method of Server to make it easier to test.

## Storing threads

As the Server creates several threads during execution, fields have been created to store these threads, to check if they were created. Could later be used to clean up running threads before the Fagi Server stops.