Server Tests
## Testing the Server class

The Server class is now tested with 100% test coverage.

### Issues

PI test is flaky around setDaemon(true) on threads. This will be ignored for now.

## Moved utility function

The Server class had a utility function to find public ip v4. This has been moved to a utility class in shared.

## RunOnceStrategy

The implementation of IsRunningStrategy that runs once in tests has been renamed from WorkerRunOnceStrategy to RunOnceStrategy. This has been done as tests of the Server class now also utilises it.

## Data

Made it possible to get the value of nextConversationId to help test loading of Conversations when starting the Server.

 ## PI test configuration

 Upgraded junit5PluginVersion from 1.2.1 to 1.2.2