Fagi Logging Framework
## Intro

We have now introduced a framework that allows the Fagi code to not care what logging framework is being used. This allows us to very easily support multiple frameworks, or completely remove the usage of a framework, without having to edit too much of the code.

## FagiLoggerFactory

This class allows for all other classes in Fagi to create a logger using the registered creation strategy. This allows us to quickly change framework a very central place.

This class also handles the configuration of the logging framework used, by calling a configuration strategy. Though it only allows calling the default configuration if no custom configuration has been provided to the framework in use.

## JavaLogger

The current logging framework to be used is the one the java standard library provides.

## Logging Config

To ensure that the logging format of the client and server are acceptable, even if the user doesn't provide a logging config properties file, we have introduced a default config.

This should only be used when there is no file specified in the property: java.util.logging.config.file.

The default config makes use of a console logger and a file logger.